ITQ4_CTF_PROJECT/app/requirements.txt
# app/requirements.txt
Flask==2.0.1
werkzeug==2.0.3
mysql-connector-python==8.0.26
